bugfix: a voicetype doesn't support emotion, resolve d.ts compile error
