+++ discord/speeches/types.d.ts
@@ -1,14 +1,5 @@
 import {v1beta1} from '@google-cloud/text-to-speech';
 
-enum Emotion {
-    normal = 'normal',
-    happiness = 'happiness',
-    anger = 'anger',
-    sadness = 'sadness',
-}
-type EmoLV = number;
-export {Emotion, EmoLV};
-
 export type SynthesizeFunction = (
 	text: string,
 	voiceType: string,
